requirements.txt
pytesseract 
pdfplumber
faiss-cpu 
sentence-transformers 
googletrans 
langdetect 
streamlit 
spacy